+++ frontend-ios/YakAl/YakAl/Sources/Views/Mypage/ConfromityTestSwiftUIView.swift
@@ -74,6 +74,7 @@ struct ConfromityTestSwiftUIView: View {
                     
                 }
             }
+
     }
 }
